+++ src/Services/Customers/ECommerce.Services.Customers/Shared/CustomersModuleConfiguration.cs
@@ -1,28 +1,14 @@
 using BuildingBlocks.Abstractions.Web.Module;
 using BuildingBlocks.Core;
 using BuildingBlocks.Monitoring;
-using ECommerce.Services.Catalogs.Shared.Extensions.ApplicationBuilderExtensions;
-using ECommerce.Services.Catalogs.Shared.Extensions.ServiceCollectionExtensions;
+using ECommerce.Services.Customers.Shared.Extensions.ApplicationBuilderExtensions;
+using ECommerce.Services.Customers.Shared.Extensions.ServiceCollectionExtensions;
 
-namespace ECommerce.Services.Catalogs;
+namespace ECommerce.Services.Customers.Shared;
 
-public class CatalogModuleConfiguration : ISharedModulesConfiguration
+public class CustomersModuleConfiguration : ISharedModulesConfiguration
 {
-    public const string CatalogModulePrefixUri = "api/v1/catalogs";
-
-    public IEndpointRouteBuilder MapSharedModuleEndpoints(IEndpointRouteBuilder endpoints)
-    {
-        endpoints.MapGet("/", (HttpContext context) =>
-        {
-            var requestId = context.Request.Headers.TryGetValue("X-Request-Id", out var requestIdHeader)
-                ? requestIdHeader.FirstOrDefault()
-                : string.Empty;
-
-            return $"Catalogs Service Apis, RequestId: {requestId}";
-        }).ExcludeFromDescription();
-
-        return endpoints;
-    }
+    public const string CustomerModulePrefixUri = "api/v1/customers";
 
     public IServiceCollection AddSharedModuleServices(
         IServiceCollection services,
@@ -49,5 +35,19 @@ public class CatalogModuleConfiguration : ISharedModulesConfiguration
         await app.UseInfrastructure(app.Logger);
 
         return app;
+    }
+
+    public IEndpointRouteBuilder MapSharedModuleEndpoints(IEndpointRouteBuilder endpoints)
+    {
+        endpoints.MapGet("/", (HttpContext context) =>
+        {
+            var requestId = context.Request.Headers.TryGetValue("X-Request-Id", out var requestIdHeader)
+                ? requestIdHeader.FirstOrDefault()
+                : string.Empty;
+
+            return $"Customers Service Apis, RequestId: {requestId}";
+        }).ExcludeFromDescription();
+
+        return endpoints;
     }
 }
